requirements.txt
streamlit
tensorflow
numpy
Pillow
gdown
streamlit-drawable-canvas
streamlit-lottie
requests